Rename examples logo components to kebab-case
