+++ packages/examples/src/components/logos/index.ts
@@ -1,4 +1,4 @@
-export * from './ExampleLogo';
-export * from './FilesLogo';
-export * from './GithubLogo';
-export * from './GoogleLogo';
+export * from './example';
+export * from './files';
+export * from './github';
+export * from './google';
